Add path_name method to classes with HTML Attachments
This commit places the knowledge of the path_name naming
within the main class rather than offloading the knowledge
to helpers or attachments.

Editions and Responses are able to have HTML Attachments.

Usually this defaults to the model name, but for Publications
with a subtype of statistics this can be specified.